refactor(player): extract DisplayHandoff from MpvController
Moves the RPi headless VT/DRM hand-off out of MpvController into a single app-owned DisplayHandoff, ahead of the work for a script module which will need the same bracket. The six helpers moved verbatim and the load-bearing sequencing moves with them rather than being duplicated:

acquire()          VT switch -> drmDropMaster -> save CRTC
releaseDeferred()  200ms -> drmSetMaster -> legacy setCrtc -> VT back

Also:
- Owner token, so two subsystems can never both hold the screen. isHeldBy() replaces the m_previousVt > 0 re-entrancy guard.
- ~MpvController now calls releaseNow(), fixing the gap where quitting mid-playback on a Pi left the VT switched away and DRM master dropped. DisplayHandoff is declared before MpvController in main() so it outlives its users.
- savedStateValid() lets a caller that can decline a hand-off do so. mpv deliberately ignores it: playback has always proceeded when the CRTC capture failed, and refusing here would change behavior.
- writeFontconfigOverride -> src/util/FontconfigOverride (fcoverride::write), so a future child mpv gets the bundled font too.
- ARCHITECTURE.md: the hand-off section named MpvController members that no longer exist; it now documents the DisplayHandoff contract.

No behavior change on the mpv path. Verified on macOS (build clean, two play/quit cycles, clean exit).  Verified on the PI to test the VT/DRM code in headless mode.
