+++ CMakeLists.txt
@@ -64,6 +64,8 @@ add_executable(240mp
     src/util/YtDlpLocator.cpp
     src/util/MpvLocator.cpp
     src/util/ExecPath.cpp
+    src/util/DisplayHandoff.cpp
+    src/util/FontconfigOverride.cpp
     src/input/InputManager.cpp
     src/input/IdleTracker.cpp
     src/update/UpdateManager.cpp
